requirements.txt
# Streamlit Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Database
sqlite3

# LangChain Dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20

# Google Generative AI
google-generativeai>=0.3.0

# SQL Database
sqlalchemy>=2.0.0

# Additional Dependencies
python-dotenv>=1.0.0